pyproject.toml
[tool.poetry]
name = "alyautdinov-ruslan-hw-12"
version = "0.1.0"
description = ""
authors = ["Аляутдинов Руслан <rusalyautdinov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
isort = "^5.13.2"
mypy = "^1.10.1"
black = "^24.4.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119

[tool.isort]
line_length = 119